chore: improve SimpleGrid exports
